# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation ",
        "description": "API Documentation",
        "contact": {
            "email": "srundavith@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/movie_genres": {
            "get": {
                "tags": [
                    "Movie_Genres"
                ],
                "summary": "Get List movie_genres Data",
                "description": "enter your movie_genres here",
                "operationId": "movie_genres",
                "responses": {
                    "default": {
                        "description": "return array model movie_genres"
                    }
                }
            },
            "post": {
                "tags": [
                    "Movie_Genres"
                ],
                "summary": "armovie_genrestists",
                "description": "movie_genres",
                "operationId": "Movie_Genres",
                "requestBody": {
                    "description": "form movie_genres",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "movie_id",
                                    "genre_id"
                                ],
                                "properties": {
                                    "movie_id": {
                                        "type": "string"
                                    },
                                    "genre_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/movie_genres/{id}": {
            "get": {
                "tags": [
                    "Movie_Genres"
                ],
                "summary": "Detail",
                "description": "-",
                "operationId": "movie_genres/GetById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "return model admin"
                    }
                }
            },
            "put": {
                "tags": [
                    "Movie_Genres"
                ],
                "summary": "Update movie_genres",
                "description": "-",
                "operationId": "movie_genres/update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "form admin",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "movie_id",
                                    "genre_id"
                                ],
                                "properties": {
                                    "movie_id": {
                                        "type": "string"
                                    },
                                    "genre_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "tags": [
                    "Movie_Genres"
                ],
                "summary": "Delete movie_genres",
                "description": "-",
                "operationId": "movie_genres/delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}